tripcnt not incrementing fix #2

+++ J1772EvseController.cpp
@@ -970,8 +970,7 @@ void J1772EVSEController::Update()
 	m_EvseState = EVSE_STATE_NO_GROUND;
 	
 	chargingOff(); // open the relay
-	
-	if ((prevevsestate != EVSE_STATE_NO_GROUND) && (m_NoGndTripCnt+1 < 254)) {
+	if ((prevevsestate != EVSE_STATE_NO_GROUND) && (((uint8_t)(m_NoGndTripCnt+1)) < 254)) {
 	  m_NoGndTripCnt++;
 	  eeprom_write_byte((uint8_t*)EOFS_NOGND_TRIP_CNT,m_NoGndTripCnt);
 	}
@@ -1021,7 +1020,7 @@ void J1772EVSEController::Update()
                ((curms - m_StuckRelayStartTimeMS) > STUCK_RELAY_DELAY) ) ||  // start delay de-bounce
 	     (prevevsestate == EVSE_STATE_STUCK_RELAY) ) { // already in error state
 	  // stuck relay
-	  if ((prevevsestate != EVSE_STATE_STUCK_RELAY) && (m_StuckRelayTripCnt+1 < 254)) {
+	  if ((prevevsestate != EVSE_STATE_STUCK_RELAY) && (((uint8_t)(m_StuckRelayTripCnt+1)) < 254)) {
 	    m_StuckRelayTripCnt++;
 	    eeprom_write_byte((uint8_t*)EOFS_STUCK_RELAY_TRIP_CNT,m_StuckRelayTripCnt);
 	  }   
@@ -1041,7 +1040,7 @@ void J1772EVSEController::Update()
     m_EvseState = EVSE_STATE_GFCI_FAULT;
 
     if (prevevsestate != EVSE_STATE_GFCI_FAULT) { // state transition
-      if (m_GfiTripCnt+1 < 254) {
+      if (((uint8_t)(m_GfiTripCnt+1)) < 254) {
 	m_GfiTripCnt++;
 	eeprom_write_byte((uint8_t*)EOFS_GFI_TRIP_CNT,m_GfiTripCnt);
       }
